# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Codespace X",
        "title": "Codespace X REST API",
        "contact": {
            "name": "Codespace Indonesia",
            "url": "https://codespace.id",
            "email": "mail@codespace.id"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "description": "List Project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "List Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization value",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ListProjectResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateTx Project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "CreateTx Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization value",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "dto.CreateProjectRequest",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateProjectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_uuid}": {
            "get": {
                "description": "Detail Project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Detail Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization value",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project_uuid",
                        "name": "project_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ListProjectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Update Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization value",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "dto.UpdateProjectReq",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProjectReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "uuid project",
                        "name": "project_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateProjectRequest": {
            "type": "object",
            "required": [
                "description",
                "service_type",
                "time_priority"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "service_type": {
                    "type": "string"
                },
                "time_priority": {
                    "type": "string"
                }
            }
        },
        "dto.CreateProjectResponse": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetProfileResponse": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "dto.ListProjectResponse": {
            "type": "object",
            "properties": {
                "astrodevs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetProfileResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "service_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail_image_url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateProjectReq": {
            "type": "object",
            "properties": {
                "deadline_type": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "service_type": {
                    "type": "integer"
                }
            }
        },
        "pkg.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/pkg.MetaResponse"
                }
            }
        },
        "pkg.MetaResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                }
            }
        }
    }
}